+++ frontend/src/components/blocks/BlockRenderer/index.js
@@ -1,0 +1,1 @@
+export { default } from './BlockRenderer'
