--- frontend/src/components/blocks/Blocks/index.js
@@ -1,1 +1,0 @@
-export { default } from './Blocks'
